requirements.txt
praw~=7.8.1
pyyaml~=6.0.2
sqlalchemy~=2.0.36
pandas~=2.2.3
vadersentiment~=3.3.2
torch~=2.5.1
numpy~=1.24.3
transformers~=4.47.1
tqdm~=4.67.1
emoji~=2.14.0
beautifulsoup4~=4.12.3
spacy~=3.7.2
pytest~=8.0.0
pytest-cov~=4.1.0
